graph, store: Remove the store-specific IdType
Use the one from graph::data::store instead
